+++ actix-rt/CHANGES.md
@@ -2,7 +2,7 @@
 
 ## Unreleased
 
-- Add `Arbiter::try_new` and `Arbiter::try_with_tokio_rt` for fallible Arbiter initialization without panicking the creator thread.
+- Add `Arbiter::try_new()` and `Arbiter::try_with_tokio_rt()` for fallible Arbiter initialization.
 
 ## 2.12.0